requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
opencv-python==4.12.0.88
Pillow==12.0.0
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.8
seaborn==0.13.2
tqdm==4.66.5
Werkzeug==3.1.4
